feat(jocaagura): Exposicion del blocCore de appManager 85.8%
